first attempt at TPA household and employment nums
this set of commits corresponds to an attempt to get the household and
employment numbers out of existing tables on the database. many of the
commits here were based around an attempt to work with and clean up
simple issues like county assignment for parcels within existing tables.

however, when reviewing the final views, it became clear that we would
also need to know about all of the parcels in TPAs that had not
seen change in employment or households.

readability-table alias

to make the query easier to read

document desired output

remove duplicate logic from main query

duplicates are handled before the main query now.

separate variables into lines

for readability

add dwelling units per acre

clean up

these queries no longer needed

duplicate checks on all scenarios

there are a bunch

resolve county duplicate spatially

there were too many to do this manually

clean up

idiom for duplicates

to remove them

resolve centroid/pos issue; add distinct stub

pos for sql server spatial does not always yield a point near the middle--added a centroid way to double check on within ness. also added a stub for how to do a distinct to drop duplicates

queries for each alternative, summaries

also, moved these down below the deduplication

add final summary views

insert doc on refactor to get 2015/2040 instead of diff values

attempt to get 2015 values out of 2040+diff

this seems like it won't work, given that the TPA and county intersections seem to be based on a selection of parcels from the footprints--therefore they could be missing a number of parcels from the TPAs that did not see change from 2015 to 2040 but which are in a TPA.