find (all) parcels in TPAs
steps to add all tpa parcels

create an indexed parcel centroid table and do an intersection on it with TPAs

create view on parcels in tpas

this is now possible more directly from existing tables and will make the final query possible.

because the view depend on having the TPA column on the parcels table,
i've moved it below the queries that achieve that.