update summary tables to include non-change parcels
draft summary table for alt 4 as spec'ed

added 2015 and 2040 units sums by county within tpa, instead of change. based on changes in previous 2 or 3 commits

summary tables for all scenarios

this was basically a lot of copy and paste

finish up nonzero views for each scenario

document script

put some headers in to clarify the long script

case statements for negative values changed

the final views in which negatives change to zeros now have to deal with change in 2015 and 2040 variables instead.